Add `HashInt` method for use with a `TypeSet` containing int IDs
Previously, `TypeList` was used for storing a collection of linked
resources. The order was not deterministic, and it would result
in unnecessary diffs. @frankh introduced a fix, but the fix did
not define a method for returning unique IDs to construct the set.
This commit will simply pass the int IDs through as unique IDs.